requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
requests==2.32.3
streamlit==1.42.2
python-dotenv==1.0.1
